[unittest] Update tests to work with refactored code
