+++ src/main/kotlin/dev/supergrecko/kllvm/ir/types/FloatType.kt
@@ -1,9 +1,9 @@
 package dev.supergrecko.kllvm.ir.types
 
 import dev.supergrecko.kllvm.internal.contracts.Unreachable
+import dev.supergrecko.kllvm.ir.Context
 import dev.supergrecko.kllvm.ir.Type
 import dev.supergrecko.kllvm.ir.TypeKind
-import dev.supergrecko.kllvm.ir.Context
 import org.bytedeco.llvm.LLVM.LLVMTypeRef
 import org.bytedeco.llvm.global.LLVM
