+++ src/main/kotlin/dev/supergrecko/kllvm/ir/values/constants/ConstantArray.kt
@@ -2,9 +2,9 @@ package dev.supergrecko.kllvm.ir.values.constants
 
 import dev.supergrecko.kllvm.internal.util.toBoolean
 import dev.supergrecko.kllvm.internal.util.toInt
+import dev.supergrecko.kllvm.ir.Context
 import dev.supergrecko.kllvm.ir.Value
 import dev.supergrecko.kllvm.ir.values.Constant
-import dev.supergrecko.kllvm.ir.Context
 import org.bytedeco.javacpp.SizeTPointer
 import org.bytedeco.llvm.LLVM.LLVMValueRef
 import org.bytedeco.llvm.global.LLVM
