+++ src/main/kotlin/dev/supergrecko/kllvm/ir/values/constants/ConstantInt.kt
@@ -3,9 +3,9 @@ package dev.supergrecko.kllvm.ir.values.constants
 import dev.supergrecko.kllvm.internal.contracts.Unreachable
 import dev.supergrecko.kllvm.internal.util.toInt
 import dev.supergrecko.kllvm.ir.Value
+import dev.supergrecko.kllvm.ir.instructions.IntPredicate
 import dev.supergrecko.kllvm.ir.types.IntType
 import dev.supergrecko.kllvm.ir.values.Constant
-import dev.supergrecko.kllvm.ir.instructions.IntPredicate
 import org.bytedeco.llvm.LLVM.LLVMValueRef
 import org.bytedeco.llvm.global.LLVM
