+++ src/main/kotlin/dev/supergrecko/kllvm/ir/values/constants/ConstantStruct.kt
@@ -1,10 +1,10 @@
 package dev.supergrecko.kllvm.ir.values.constants
 
 import dev.supergrecko.kllvm.internal.util.toInt
+import dev.supergrecko.kllvm.ir.Context
 import dev.supergrecko.kllvm.ir.Value
 import dev.supergrecko.kllvm.ir.types.StructType
 import dev.supergrecko.kllvm.ir.values.Constant
-import dev.supergrecko.kllvm.ir.Context
 import org.bytedeco.javacpp.PointerPointer
 import org.bytedeco.llvm.LLVM.LLVMValueRef
 import org.bytedeco.llvm.global.LLVM
