+++ src/main/kotlin/dev/supergrecko/kllvm/support/MemoryBuffer.kt
@@ -2,8 +2,8 @@ package dev.supergrecko.kllvm.support
 
 import dev.supergrecko.kllvm.internal.contracts.Disposable
 import dev.supergrecko.kllvm.internal.contracts.Validatable
-import dev.supergrecko.kllvm.ir.Module
 import dev.supergrecko.kllvm.ir.Context
+import dev.supergrecko.kllvm.ir.Module
 import org.bytedeco.llvm.LLVM.LLVMMemoryBufferRef
 import org.bytedeco.llvm.LLVM.LLVMModuleRef
 import org.bytedeco.llvm.global.LLVM
