+++ src/test/kotlin/dev/supergrecko/kllvm/ir/BuilderTest.kt
@@ -1,16 +1,13 @@
-package dev.supergrecko.kllvm.builders
+package dev.supergrecko.kllvm.ir
 
-import dev.supergrecko.kllvm.ir.Value
 import dev.supergrecko.kllvm.ir.types.FunctionType
 import dev.supergrecko.kllvm.ir.types.IntType
 import dev.supergrecko.kllvm.ir.types.VoidType
 import dev.supergrecko.kllvm.ir.values.constants.ConstantInt
-import dev.supergrecko.kllvm.ir.Builder
-import dev.supergrecko.kllvm.ir.Module
+import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertNull
-import org.junit.jupiter.api.Test
 
 class BuilderTest {
     @Test
