+++ src/test/kotlin/dev/supergrecko/kllvm/ir/types/TypeTest.kt
@@ -1,10 +1,10 @@
 package dev.supergrecko.kllvm.ir.types
 
-import dev.supergrecko.kllvm.ir.TypeKind
 import dev.supergrecko.kllvm.ir.Context
+import dev.supergrecko.kllvm.ir.TypeKind
+import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
-import org.junit.jupiter.api.Test
 
 class TypeTest {
     @Test
