+++ src/test/kotlin/dev/supergrecko/kllvm/ir/values/ConstantIntTest.kt
@@ -2,8 +2,8 @@ package dev.supergrecko.kllvm.ir.values
 
 import dev.supergrecko.kllvm.ir.types.IntType
 import dev.supergrecko.kllvm.ir.values.constants.ConstantInt
-import kotlin.test.assertEquals
 import org.junit.jupiter.api.Test
+import kotlin.test.assertEquals
 
 class ConstantIntTest {
     @Test
